+++ classes/class-base.php
@@ -13,7 +13,7 @@ namespace Progress_Planner;
  * @method \Progress_Planner\Suggested_Tasks get_suggested_tasks()
  * @method \Progress_Planner\Settings get_settings()
  * @method \Progress_Planner\Query get_query()
- * @method \Progress_Planner\Cache get_cache()
+ * @method \Progress_Planner\Utils\Cache get_utils__cache()
  * @method \Progress_Planner\Page_Types get_page_types()
  * @method \Progress_Planner\Rest_API_Stats get_rest_api_stats()
  * @method \Progress_Planner\Rest_API_Tasks get_rest_api_tasks()
