+++ classes/class-page-types.php
@@ -89,7 +89,7 @@ class Page_Types {
 	 * @return void
 	 */
 	public function maybe_update_terms() {
-		$updated = \progress_planner()->get_cache()->get( 'page_types_updated' );
+		$updated = \progress_planner()->get_utils__cache()->get( 'page_types_updated' );
 		if ( $updated ) {
 			return;
 		}
@@ -136,7 +136,7 @@ class Page_Types {
 			}
 		}
 
-		\progress_planner()->get_cache()->set( 'page_types_updated', true, \WEEK_IN_SECONDS );
+		\progress_planner()->get_utils__cache()->set( 'page_types_updated', true, \WEEK_IN_SECONDS );
 	}
 
 	/**
